Create TechNumber.java
This is a program to test if a user given number is a Tech number or not. What is a Tech number? A Tech number is a number that has an even number of digits, and if the number is split into two equal halves (from the middle), then the square of the sum of these halves is equal to the number itself.
For example, consider the number 81. When divided equally, we have 8 and 1. Now 8+1=9 and 9^2=81. Hence, 81 is a Tech number.