Create ISBNnumber13.java
To validate an ISBN-13 number, ensure it has exactly 13 digits and that it satisfies the ISBN-13 checksum formula. The checksum formula for ISBN-13 is calculated as follows:

>Multiply each of the first 12 digits by a weight, alternating 
between 1 and 3, starting with 1.
>Sum all these values.
>The checksum digit (the 13th digit) is the number that must be 
added to the sum to make it divisible by 10. In other words, it's 
the value that makes the sum a multiple of 10.

For example: 8902519002228
8*1 + 9*3 + 0*1 + 2*3 + 5*1 + 1*3 + 9*1 + 0*3 + 0*1 + 2*3 + 2*1 + 2*3 = 72
72+8 = 80 
80 is multiple of 10
Therefore it is a valid 13 digit ISBN number