Create ToggleCase.java
To replace lower-case characters with upper-case and vice-versa in a string in Java, you can iterate through the characters of the string, check the case of each character, and convert it to the opposite case. This can be done using the Character.isLowerCase() and Character.isUpperCase() methods along with Character.toLowerCase() and Character.toUpperCase().