Create Permutations.java
Tony Stark is teaching Jarvis different algorithms and to test what Jarvis learned he asked many questions.Out of which one question is to print all permutations of a given string. Write a program to print all permutations of a string in lexicographical order. String A is lexicographically smaller than string B if A[j] < B[j] , 1 <= j <= N

and A[i] = B[i] for all 1 <= i < j